document-assistant/backend/requirements.txt
fastapi
uvicorn
langchain-groq
langchain-community
langchain-core
huggingface-hub
faiss-cpu
pydantic
python-dotenv